Remove support for at-cuda calling precompiled CuFunctions.
This was sketchy anyway, as we didn't know the types the function was compiled for.
But worse, it was broken due to (what's probably a bug or yet another disallowed use
of) the generated function behind at-cuda: during the inference-stage execution,
it always dispatched to the implementation for F<:CuFunction, even though the argument
most definitely isn't a CuFunction (and runtime dispatch worked fine).

Due to lack of time to debug and upstream (and get an EWONTFIX) this, I've removed the code path.