improved mapreducedim kernel
The algorithm here is to distribute 1024 threads/block among the inner
(reduction) and outer (map) dimensions; the remaining outer dimensions
get scheduled as the block grid. If fewer than 8 threads would be
dedicated to the inner dimension (because the outer dimension is much
larger) then we revert to the original kernel (which has the advantage
that it doesn't need to copy data to and from shared memory).