Infrastructure for multiple memory pools (#414)
Mostly NFC, but changes internal APIs and includes fixes to the binned memory pool,
the old implementation that is still used by default. Specifically, two issues, (a) the largest
pool would be inefficient (partially overlapping with the MAX_POOL cutoff) but
also potentially freeing pointers from it according to the large_alloc regime,
but worse (b) views would report the size of their allocation and thus size of
the object being freed as the size of the slice instead of the full array,
leading to potentially buffer being put in the wrong pool.