ND-reverse (#416)
* added reverse for arbitrary dimension CuArrays

* removed a redundant variable

* Simplified kernel

* Whitespace fixes.

* More compact testing.

* Clean-up reverse kernels.

* Drop specific typing.

* Unify kernels.

* Working, but non-coalesced in-place ND.

* Make the ND-case coalesced for 1D inputs.

* Unify in-place and out-of-place kernels.

Using shared memory for out-of-place is better since that enables memory coalescing on the output write.

* Drop the shared memory, use separate inplace/outofplace kernels.

Wrt. shared memory: newer hardware can coalesce both forward and backward access patterns,
which we would get without shared memory in the 1D case, while for the ND case we don't
get either of those if just writing backwards in the shared memory buffer.

Wrt. separate kernels: the in-place version would trample over values to be read by other blocks.
Grid synchronization is much to costly, so just swap numbers on half the number of threads.