Re-initialize the random seed every time.
Turns out we can't rely on the shared memory persisting across launches,
so we need to initialize the key at every launch. That also implies we
need a fresh seed for every launch, because otherwise two subsequent
launches would generate the same numbers. To avoid having to upload a
seed before every launch, use the `clock` register as a source of
randomness to initialize the key with.