Don't assume host pointers are directly usable on the device. (#1342)
Since CUDA.jl 3.4, we now always reconstruct a buffer object when
creating an array from a pointer. In the case of a HostBuffer,
that means we want the host pointer in the buffer object,
which is converted to a device pointer on request.

However, unsafe_wrap is invoked with a device pointer. Generally,
those pointers are identical, but on GPUs where
CU_DEVICE_ATTRIBUTE_CAN_USE_HOST_POINTER_FOR_REGISTERED_MEM==0,
that is not the case.

Instead, recover the host pointer from the device pointer passed to
unsafe_wrap. This is going back and forth between both pointer
representations, but those calls are reasonably cheap.