CUTENSOR: axpy! and axpby! not mutating fixed (#1416)
* In-place multiplication

axpy! and axpby! are supposed to overwrite the input (in-place multiplication) without allocating a new CuTensor in memory. This fixes the incorrect behavior as well as improving performance.

* Tests for in-place behavior

Creates a copy of input tensor before calling axpy! and axpby! and checks if the returned tensor points to the overwritten input.

* Refractor by implementing CuTensor copying

Implement Base.copy() for CuTensors which duplicates the object with all its contents.