rand: seed kernels from the host.
We should not (and cannot) rely on the fact that updating shared memory
will be visible across kernel launches. In some cases, it isn't, which
combined with pre-initialized shared memory (resulting in us not hitting
the make_seed path) causes identical values for the key and counters,
resulting in identical random numbers getting generated.

Avoid this by passing a seed from the host. Even in the case that updates
to counters don't survive to the next kernel, now the key will be different,
resulting in new numbers being generated.