docs: perf tips: deemphasize `assume` in favor of UnsafeAssume.jl
The new package is dedicated to functionality like `assume`. It has
documentation and examples, and the functions have better effects as
appropriate depending on the Julia version. For example, the new
package preserves a `nothrow` effect of the calling code. Unreleased
versions of Julia are able to eliminate `try`-`catch` constructs when
`nothrow` is known.