Refactor memory wrappers (#2335)
This commit significantly reworks the CUDA.jl memory wrappers.

The low-level driver wrappers have been moved out of the Mem submodule,
and for consistency with Base the Buffer name has been replaced by
Memory. This is mostly an NFC change, and deprecations have been added
to make this non-breaking.

The high-level wrappers have been changed to to introduce a Managed
layer, i.e., CuArray now container a `DataRef` (from GPUArrays, for
refcounting), which contains a new `Managed`, which contains an
`AbstractMemory` subtype. The purpose of the `Managed` struct is to keep
track of the state of the memory, specifically, which stream has last
modified it, and whether there's outstanding asynchronous operations.
This information is used synchronize the memory when deriving a pointer,
making it possible to avoid explicit synchronization in many cases
(notably, when switching tasks or devices).

Sadly, this approach is not water-tight, notably, when memory is
captured (e.g. as part of a CUDA graph) we may not hit the pointer
derivation path and thus fail to synchronize. Explicit synchronization
is still required here, or an explicit copy to CPU memory (in which case
CUDA.jl will always synchronize).

As part of this change, several bugs and other minor issues were
discovered and fixed. Refer to the GitHub PR for an un-squashed version
of these changes for more details.