Handle cache improvements (#2352)
Rework the handle cache so that handle ctors are dtors are known to the
cache. This makes it possible for it to evict handles at any time, and is
used to wipe handles when reclaiming memory.

Also switch to a spinlock to ensure safe operation in finalizers,
and consider collecting memory when allocating new handles.